+++ VSCode/branchnamegenerator/CHANGELOG.md
@@ -7,3 +7,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [Unreleased]
 
 - Initial release
+
+### [1.0.8]
+
+- Branch name is copied to clipboard when creating a new branch. 
+- Will now ask if you would like to checkout the created branch.